test: consolidate API-key runtime setup
Use a per-test isolated auth baseline, parameterized access-resolution cases, and a compact extension runtime helper. Keep representative API-key lifecycle and terms-notice coverage without duplicated fetch-order scaffolding.
